Puntos finales

+++ src/tienda/tienda.service.ts
@@ -1,9 +1,9 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateTiendaDto } from './dto/create-tienda.dto';
 import { UpdateTiendaDto } from './dto/update-tienda.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { tienda } from './entities/tienda.entity';
-import { Repository } from 'typeorm';
+import { In, Repository } from 'typeorm';
 
 @Injectable()
 export class TiendaService 
@@ -19,15 +19,26 @@ export class TiendaService
   }
 
   async findAll() {
-    return await this.tiendaRepository.find();
+    const tiendas = await this.tiendaRepository.find();
+    if (!tiendas) {
+      throw new NotFoundException(`No se encontraron Tiendas`);
+    }
+    return tiendas;
   }
 
   async findOne(id: number) {
-    return await this.tiendaRepository.findOne({where:{id:id}});
+    const producto= await this.tiendaRepository.findOne({where:{id:id}});
+    if (!producto) {
+      throw new NotFoundException(`No se encontró la Tienda ${id}`);
+    }
+    return producto;
   }
 
   async update(id: number, updateTiendaDto: UpdateTiendaDto) {
     const tienda=await this.tiendaRepository.findOne({where:{id:id}});
+    if (!tienda) {
+      throw new NotFoundException(`No se encontró la Tienda ${id}`);}
+
     const nuevatienda={
       ...tienda,
       ...updateTiendaDto
@@ -37,7 +48,16 @@ export class TiendaService
 
   async remove(id: number) {
     const tienda=await this.tiendaRepository.findOne({where:{id:id}});
+    if (!tienda) {
+      throw new NotFoundException(`No se encontró la Tienda ${id}`);
+    }
     await this.tiendaRepository.delete(tienda)
     return "Tienda Eliminada";
+  }
+
+  async findIn(tiendasId:number[]){
+    return await this.tiendaRepository.find({
+      where: { id: In(tiendasId) }
+    })
   }
 }
